pyproject.toml
[project]
name = "sismon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "google-genai>=1.51.0",
    "librosa>=0.11.0",
    "moviepy>=2.2.1",
    "numpy>=2.3.5",
    "python-dotenv>=1.2.1",
    "scipy>=1.16.3",
    "soundfile>=0.13.1",
    "streamlit>=1.51.0",
]
